+++ Changes.md
@@ -14,6 +14,7 @@ Fixes
 -----
 
 - Editor : Fixed `Internal C++ object already deleted` errors when some editors were destroyed.
+- UVInspector : Fixed `Unable to find ScriptNode for UVView` warnings.
 
 API
 ---